xhci: support USB_HOOK_DISCARD
"usb_hook_process()" returns an "int" value (e.g., USB_HOOK_DISCARD,
which is 0x20000), but the xHCI driver previously stored the returned
value in a "u8" variable. As a result, the driver could not handle the
variable as expected.

This commit changes the driver to store the returned value in an "int"
variable. With this commit, the driver can handle the variable
returned from "usb_hook_process()" correctly.

+++ drivers/usb/xhci.c
@@ -1131,9 +1131,9 @@ handle_slot_write (struct xhci_host *host, uint slot_id, uint ep_no)
 	while (g_urb) {
 		struct usb_request_block *h_urb = xhci_shadow_g_urb (g_urb);
 
-		u8 ret = usb_hook_process (host->usb_host,
-					   h_urb,
-					   USB_HOOK_REQUEST);
+		int ret = usb_hook_process (host->usb_host,
+					    h_urb,
+					    USB_HOOK_REQUEST);
 
 		if (ret == USB_HOOK_DISCARD) {
 			dprintft (1, "slot: %u ep_no: %u Request discarded\n",
